requirements.txt
aiogram==2.16
django==3.2.20
google-auth==2.23.0
google-api-python-client==3.4
google-auth-oauthlib
google-auth-httplib2==0.1.0
google-api-core==2.1.1
httplib2==0.20.2
requests==2.26.0
aiogram==2.16
logging
python-dotenv
oauth2client